Have now got construction of the SziTileSource working, and have tested it.
Because OSD assumes you are in the browser, I've used jsdom to emulate this,
but to my surprise it messes up URL.toString(), which is why there are some
surprising url.hrefs kicking around.

Somewhat more excitingly, this proves that we can fetch the config dzi file
from the server and configure the base class with it. I now 'just' need to
override the image loading and we should have something working.

Have also tweaked the range function so that its end bound is exclusive as
I don't want to have to subtract one all over the place.