Got ZIP64 working
The code's a wee bit messier now, but the initial good news is that for VIPS
generated files at least, it doesn't seem like there is too much junk that
might get in the way of a more forwards compatible version of the implementation.

One limitation is that I have decided not to bother with bigints, as it was
getting very messy when I tried to use them. So any files bigger than
2^53 bytes won't work. Luckily, I don't think we're going to be looking at files
bigger than 8 pebibytes for the forseeable.

I've added a bit more bounds checking as I've gone, but I really should be
actively checking magic numbers and other inconsistencies as well, as it will
make figuring out why a given file won't load easier.