Now works with manually zipped szi files
I ran into some bother with extra fields having different lengths in the local
and central director headers. The only reliable way to do deal with this is to
actually read and parse the headers of the individual files when loading them,
so that's what I did. It does necessitate fetching everthing from the start of
one entry in the the zip to the start of the next in each request, but this
is easy enough to do with a bit of fiddling about.