Add support for download request options and abort
Rather than using the existing OSD options for Ajax requests, which are geared
towards calling XmlHttpRequest, I instead implemented the ability to set extra
headers, cors mode, and whether/how to pass credentials on a request using a
fetchOptions object that mirrors those aspects of a modern Fetch API exactly.
The only slight quirk is that mode:'no-cors' errors out, as setting it on a
request means Range headers are ignored!

The abort implementation is straightforward, using the standard AbortController
and signal mechanism.

Doing all this meant adding some more docs, which in turn made me realise that
the boundaries between a few bits of code weren't in the optimal place, so
there's been a wee bit of shuffling around and tidying up as I've gone. I also
removed the on document load stuff in index.html because it simply wasn't
needed.