Split main.js into provisional modules
Doing this to help comprehensibility and as a prelude to getting a proper
build for distribution. This splits the code into the the obvious four modules,
each centered around an existing class. I think it might also make sense to split
the sziFileReader into two parts - one that's just for dealing with zip parsing
and one for what's left - but it's not clear to me exactly where the line is!