Added build config, so you can now build for distribution
This was suprisingly easy. It's mostly just set to defaults now, though I did
have to do some mucking around to get it to exclude the public directory from
the build files. I think we might want to organise the folder structure a bit
differently, but that can wait for now...