Test LittleEndianDataReader, clarify string encoding, fix bounds checks
While thinking about adding the tests, I realised that I wasn't totally sure
that we can rely on the filenames being UTF-8 encoded. It turns out that while
most of the time they will be, it's not guaranteed, but in practice it doesn't
matter. I left a big comment explaining this at the point where we first read
a filename, and renamed the data reader function to readUtf8String, to make it
as clear as possible what is being read and how.

When actually writing them, I uncovered a typo in the bounds checking code that
meant you could overrun a buffer. Luckily it's trivial to fix.