Happy path tests for sziFileReader
This was mostly straightforward to add - just reading in the existing example
files and checking they were readable. The image file checks are fairly loose
but all validate that the start and end of the files are where we expect them
to be, which is the main thing that can go wrong when reading the file.

The one faffy thing was creating a non-huge ZIP64 file. Python's ZipFile lib
claims to support forcing ZIP64 but it only seems to work for local headers,
so I had to do some hacking at the base class in a bit of my own code to make
it work.