Simplified the generation of the table of contents
When tidying the code up I realised that we were doing an unnecessary step in
generating a map from start to ends and then creating the the map from
filenames to start to end. This made things more obscure than they needed to be.
So instead, we now just sort the entire central directory backwards and iterate
through it backwards, using the start of the previous file data as the end of
the next.