Add handling for unexpected magic number in EOCD comment
Honestly, this is pretty unlikely to occur as the magic number contains
two ASCII control chars but it was bugging me that we weren't trying
to deal gracefully with it. All the handling does is check that the
EOCD starting at the found magic number reaches the end of the file
exactly, and if it doesn't, continues the search. This isn't
bombproof, but it takes a very unlikely occurrence and makes it a lot
less likely.