Replace use of dev server with msw for network dependent tests
This means `pnpm test` should work as expected out of the box, without
the need for anyone to read the README.md to figure out why the tests
aren't working. As part of doing this, I also expanded what is being
tested a tiny bit and fixed a bit of of buggy error handling.