Allow SZI files without a single top level folder
According to the spec, the SZI file should contain a single top level
folder with the same name as the .dzi file minus its extension. In
practice, it's quite easy to miss this requirement, and zip up the dzi
file and the image tile pyramid from inside their common folder. This
change means that files generated that way will work with the Tile
Source.