Use Content-Range rather than Content-Length to get file length (#2)
The old method of making a HEAD request to the SZI file and assuming
that the returned Content-Length header turned out to be
unreliable. Browsers generally set the Accept-Encoding to allow
compressed responses and servers that support compression will favour
it, all other things being equal. This means that in the general case
the Content-Length from the HEAD could well be the *compressed* length
of the file, *not* the real length.

So, for files stored on some systems (notably Github pages), we ended up
with the wrong chunk of data being returned when looking for the Central
Directory, meaning that the code couldn't find any of the expected data,
which caused it to error out.

It turns out that you can resolve this outside the browser by setting
the Accept-Encoding header yourself, but inside the browser this is a
"Forbidden Header", which can't be overridden, so this isn't an option
for us.

Luckily, if you do a ranged GET request for a small fragment of the file
*most* browsers set Accept-Encoding to "identity", which effectively
forbids compression, and *most* servers will respond to Ranged requests
with uncompressed responses containing a handy Content-Range header,
whose value contains the total file length. This commit changes the code
to use this mechanism instead, which should work in most situations.

However, there are still combinations of browser and server that won't
work. Firefox doesn't set Accept-Encoding to "identity" for ranged
requests, and the Github pages server responds with compressed response
bodies even when faced with a ranged request. There doesn't seem to be
anything reasonable we can do to fix this, other than warn folks that
it's a possibility.